requirements.txt

python-dotenv==0.13.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pycparser==2.20
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1